+++ src/main/java/com/example/demo/client_ui/dto/order/OrderBriefDTO.java
@@ -15,7 +15,7 @@ public class OrderBriefDTO {
 
     private Integer numberOfItems;
 
-    private Float totalPrice;
+    private Integer totalPrice;
 
     private String status;
 }
